[SW-6282] Remove support for Project and Cohort metrics (#56)
Making a number of breaking changes to Project and Cohort resource
classes

* Metrics have been deprecated: they are also being removed from the
underlying API.
* Removed the `CohortPatientMetadata[Set]` classes: they serve the same
function as `ProjectPatientMetadata[Set]`, and there's no reason to
maintain both. I considered leaving it in place, and just making it an
alias of the project class.... but I don't think this part of the SDK is
heavily used, and I figured I'd clean it up while I'm making other
breaking changes. LMK if you disagree!
* The `code_name` attribute on the ProjectPatientMetadata object has
been renamed to `project_code_name`, for clarity. The column in the
dataframe has also been renamed.