[SW-7024] Fix warnings in runeq (#62)
[Jira ticket](https://runelabs.atlassian.net/browse/SW-7024) 

This fixes the warning `FutureWarning: The behavior of DataFrame
concatenation with empty or all-NA entries is deprecated. In a future
version, this will no longer exclude empty or all-NA columns when
determining the result dtypes. To retain the old behavior, exclude the
relevant entries before the concat operation.` from coming up when
retrieving stream data with the get datastream runeq function.