feat: item set generics (#80)
When working with concrete implementations of ItemSet, type checkers and
LSPs like pyright/pylance, mypy, etc. see the result of a `get` or other
operations as being of instance `ItemBase`. This is accurate but it
means that accessing properties declared on a subclass of ItemBase is
flagged as a type error, even though the item we got back was of that
subclass.

For example, suppose we have the following code:

```python
from typing import cast
from runeq.resources.stream_metadata import get_patient_stream_metadata, StreamMetadata, ItemBase

stream_metadata_set = get_patient_stream_metadata(patient_id)

for stream_metadata in stream_metadata_set:
    print(isinstance(stream_metadata, StreamMetadata))  # True
    print(isinstance(stream_metadata, ItemBase))  # True
    category = stream_metadata.parameters.get("category")
    #                          ~~~~~~~~~~
    # Pyright: Cannot access attribute "parameters" for class "ItemBase"
    #   Attribute "parameters" is unknown
    stream_metadata = cast(StreamMetadata, stream_metadata)
    category = stream_metadata.parameters.get("category")  # All good
```
At run time, there is no issue. But while developing with the library,
it's handy for the type checker to know what's up.

This PR adds generics to the `ItemSet` base to enable type checkers to
know the concrete type of the items in the set.

Signed-off-by: Daniel Bluhm <dbluhm@runelabs.io>